Move the kernel header file to a separate include path
No functional changes, it just keeps the code more obviously separate. And we
can drop the various checks for the installed header now.
