+++ configure.ac
@@ -35,7 +35,6 @@ AC_CHECK_PROG(HAVE_DOXYGEN, [doxygen], [yes], [no])
 AM_CONDITIONAL(HAVE_DOXYGEN, test "x$HAVE_DOXYGEN" = xyes)
 
 PKG_CHECK_MODULES(GLIB, glib-2.0 gudev-1.0)
-AC_CHECK_HEADERS(linux/input.h)
 
 PKG_CHECK_MODULES(LIBXML, libxml-2.0 glib-2.0, HAVE_LIBXML="yes", HAVE_LIBXML="no")
 AM_CONDITIONAL(HAVE_LIBXML, test x$HAVE_LIBXML = xyes)
